requirements.txt
openai
pandas
numpy
scipy
statsmodels
tqdm
python-dotenv
matplotlib
requests
jsonlines
ollama
regex
